requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers
faster-whisper
pytest
openai
python-dotenv
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz